tests/wpt/css-tests/css-transforms-1_dev/xhtml1print/2d-rotate-js.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
		<title>JS test: Rotate via javascript must show the correct computed rotation</title>
		<link href="http://mrkn.co/axegs" rel="author" title="Rick Hurst" />
		<link href="http://www.w3.org/TR/css-transforms-1/#transform-property" rel="help" />
		<script src="/resources/testharness.js"></script>
		<script src="/resources/testharnessreport.js"></script>
		<meta content="svg" name="flags" />
		<meta content="Asserting that you can rotate an element with JS and it show up in CSS computed values not as a matrix but as the rotation" name="assert" />
		<style>
			#box{
				margin-top:30px;
				display: block;
				width: 50px;
				height: 50px;
				background-color: green;
			}
		</style>
	</head>
    <body>
		<h1>Rotate via JS</h1>
        <div id="log"></div>
        <div id="box"></div>
		<script>
			test(function() {
				var box = document.getElementById("box");
				box.style.transform = "rotate(30deg)";
				assert_equals(window.getComputedStyle(box).getPropertyValue("transform"),
				              "rotate(30deg)");
			});
		</script>
    

</body></html>
